% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt-methods.R
\name{ppt_logLik}
\alias{ppt_logLik}
\alias{logLik.ppt}
\title{Extract the posterior mean point-process log likelihood}
\usage{
ppt_logLik(object, ...)

\method{logLik}{ppt}(object, ...)
}
\arguments{
\item{object}{A fitted `ppt` object.}

\item{...}{Reserved for compatibility.}
}
\value{
An object of class `logLik`.
}
\description{
This is the posterior mean in-sample log likelihood, not model evidence.
}
