% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt-methods.R
\name{ppt_summary}
\alias{ppt_summary}
\alias{summary.ppt}
\title{Summarize a fitted PPT}
\usage{
ppt_summary(object, ...)

\method{summary}{ppt}(object, ...)
}
\arguments{
\item{object}{A fitted `ppt` object.}

\item{...}{Reserved for future methods.}
}
\value{
An object of class `summary.ppt`.
}
\description{
Summarize a fitted PPT
}
