% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt-methods.R
\name{ppt_lppd}
\alias{ppt_lppd}
\title{Extract test-pattern log predictive density}
\usage{
ppt_lppd(object)
}
\arguments{
\item{object}{A fitted `ppt` object.}
}
\value{
A numeric scalar.
}
\description{
`test` must be supplied to `ppt_fit()`. The returned value is the joint
posterior log predictive density of that test point pattern over the fitted
observation region.
}
